# Improve indexing performance: batch stat()s with io_uring