--- a/lib/tests/meson.build
+++ b/lib/tests/meson.build
@@ -86,7 +86,7 @@ test('test-scanner',
                 install: false,
                 cpp_args: ['-DBUILD_TESTS'],
                 dependencies: [glib_dep, config_h_dep,
-                               lib_mu_utils_dep]))
+                               lib_mu_dep]))
 
 test('test-xapian-db',
      executable('test-xapian-db',