--- a/lib/utils/meson.build
+++ b/lib/utils/meson.build
@@ -34,16 +34,22 @@ else
   test_srcs =  []
 endif
 
-lib_mu_utils=static_library('mu-utils',
- [ srcs, test_srcs ], dependencies: [
+lib_mu_utils_dep_list= [
   glib_dep,
   gio_dep,
   gio_unix_dep,
   config_h_dep,
   readline_dep,
   cld2_dep
-], include_directories:
-     include_directories(['.', '..', thirdparty]),
+]
+
+if liburing_dep.found()
+  lib_mu_utils_dep_list += liburing_dep
+endif
+
+lib_mu_utils=static_library('mu-utils',
+ [ srcs, test_srcs ], dependencies: lib_mu_utils_dep_list,
+   include_directories: include_directories(['.', '..', thirdparty]),
 install: false)
 
 lib_mu_utils_dep = declare_dependency(