--- a/meson_options.txt
+++ b/meson_options.txt
@@ -61,3 +61,8 @@ option('readline',
        type: 'feature',
        value: 'auto',
        description: 'enable readline support for the mu4e repl')
+
+option('iouring',
+       type: 'feature',
+       value: 'auto',
+       description: 'enable io_uring support for faster file operations')