@djcb 2025-02-23
Looks interesting. Will take a bit longer to review this (I'm not too familiar with io_uring), I don't think before 1.12.9 (which should be in the coming week)

The performance improvements are modest, so we'll have to weigh that against the added complexity. Anyway, thanks for doing this!